refactor(client): HASHI-105 예약 취소 후 캐시 동기화 로직을 reservation feature로 승격
