+++ apps/client/src/features/reservation/index.ts
@@ -1,2 +1,14 @@
 export { cancelReservation } from './api/cancelReservation'
+export { getMyReservations } from './api/getMyReservations'
+export type {
+  MyReservationsApiStatus,
+  ReservationListResponse,
+  ReservationResponse,
+} from './api/getMyReservations'
 export { useCancelReservationMutation } from './hooks/useCancelReservationMutation'
+export { myReservationsQueryKeys } from './queries/myReservationsQueryKeys'
+export { syncCanceledReservationCache } from './queries/syncCanceledReservationCache'
+export {
+  myReservationsInfiniteQueryOptions,
+  useMyReservationsInfiniteQuery,
+} from './queries/useMyReservationsInfiniteQuery'
